add faa_notam table and change elevation column type to double

+++ navplan_backend/php-app/src/Navplan/Aerodrome/Persistence/Model/DbTableAirport.php
@@ -34,7 +34,7 @@ class DbTableAirport extends DbTable
         $this->addCol(self::COL_ICAO, DbColType::STRING, true);
         $this->addCol(self::COL_LATITUDE, DbColType::DOUBLE);
         $this->addCol(self::COL_LONGITUDE, DbColType::DOUBLE);
-        $this->addCol(self::COL_ELEVATION, DbColType::INT);
+        $this->addCol(self::COL_ELEVATION, DbColType::DOUBLE);
         $this->addCol(self::COL_ZOOMMIN, DbColType::INT, true);
         $this->addCol(self::COL_GEOHASH, DbColType::STRING);
         $this->addCol(self::COL_LONLAT, DbColType::GEO_POINT);
